@@ -5,27 +5,26 @@
     "keywords": ["framework", "laravel", "api"],
     "license": "MIT",
     "require": {
-        "php": "^8.2",
+        "php": "^8.3",
         "bugsnag/bugsnag-laravel": "^2.28",
         "guzzlehttp/guzzle": "^7.8.1",
-        "laravel/framework": "^10.48.15",
-        "laravel/sanctum": "^3.3.3",
-        "laravel/tinker": "^2.9.0",
+        "laravel/framework": "^11.14",
+        "laravel/sanctum": "^4.0.2",
+        "laravel/tinker": "^2.9",
         "predis/predis": "^2.2.2",
-        "promphp/prometheus_client_php": "^2.10.0",
+        "promphp/prometheus_client_php": "^2.10",
         "treblle/treblle-laravel": "^4.5.10"
     },
     "require-dev": {
-        "barryvdh/laravel-ide-helper": "^2.15.1",
         "fakerphp/faker": "^1.23.1",
-        "larastan/larastan": "^2.9.7",
+        "larastan/larastan": "^2.9.8",
         "laravel/pint": "^1.16.1",
         "mockery/mockery": "^1.6.12",
-        "nunomaduro/collision": "^7.10.0",
+        "nunomaduro/collision": "^8.1.1",
         "nunomaduro/phpinsights": "^2.11",
         "pestphp/pest": "^2.34.8",
         "pestphp/pest-plugin-laravel": "^2.4",
-        "rector/rector": "^0.16.0"
+        "rector/rector": "^1.2"
     },
     "autoload": {
         "psr-4": {
@@ -51,7 +50,9 @@
             "@php -r \"file_exists('.env') || copy('.env.example', '.env');\""
         ],
         "post-create-project-cmd": [
-            "@php artisan key:generate --ansi"
+            "@php artisan key:generate --ansi",
+            "@php -r \"file_exists('database/database.sqlite') || touch('database/database.sqlite');\"",
+            "@php artisan migrate --graceful --ansi"
         ],
         "post-install-cmd": [
             "./dev/hooks/install"
