@@ -11,17 +11,17 @@ use Throwable;
 
 class PreCommitCommand extends Command
 {
-    protected const GIT_DIFF_COMMAND = 'git diff --diff-filter=ACMR --name-only --cached  | grep \'\.php$\'';
+    protected const string GIT_DIFF_COMMAND = 'git diff --diff-filter=ACMR --name-only --cached  | grep \'\.php$\'';
 
-    protected const GIT_ADD_COMMAND = 'git add %s';
+    protected const string GIT_ADD_COMMAND = 'git add %s';
 
-    protected const PHP_LINT_COMMAND = 'php -l %s';
+    protected const string PHP_LINT_COMMAND = 'php -l %s';
 
-    protected const PINT_COMMAND = '%s/vendor/bin/pint %s';
+    protected const string PINT_COMMAND = '%s/vendor/bin/pint %s';
 
-    protected const PHP_STAN_COMMAND = '%s/vendor/bin/phpstan analyse %s --memory-limit=256M';
+    protected const string PHP_STAN_COMMAND = '%s/vendor/bin/phpstan analyse %s --memory-limit=256M';
 
-    protected const PEST_COMMAND = './vendor/bin/pest --parallel';
+    protected const string PEST_COMMAND = './vendor/bin/pest --parallel';
 
     /**
      * The name and signature of the console command.
