@@ -6,6 +6,4 @@ namespace App\Exceptions;
 
 use Symfony\Component\HttpKernel\Exception\HttpException;
 
-class ApiException extends HttpException
-{
-}
+class ApiException extends HttpException {}
