@@ -4,12 +4,9 @@ declare(strict_types=1);
 
 namespace App\Http\Controllers;
 
-use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
-use Illuminate\Foundation\Validation\ValidatesRequests;
-use Illuminate\Routing\Controller as BaseController;
+use App\Traits\HasApiResponse;
 
-class Controller extends BaseController
+abstract class Controller
 {
-    use AuthorizesRequests;
-    use ValidatesRequests;
+    use HasApiResponse;
 }
