@@ -81,9 +81,7 @@ return [
     |
     */
 
-    'filters' => empty(env('BUGSNAG_FILTERS'))
-        ? null
-        : explode(',', str_replace(' ', '', strval(env('BUGSNAG_FILTERS')))),
+    'filters' => empty(env('BUGSNAG_FILTERS')) ? null : explode(',', str_replace(' ', '', env('BUGSNAG_FILTERS'))),
 
     /*
     |--------------------------------------------------------------------------
@@ -110,9 +108,7 @@ return [
     'proxy' => array_filter([
         'http' => env('HTTP_PROXY'),
         'https' => env('HTTPS_PROXY'),
-        'no' => empty(env('NO_PROXY'))
-            ? null
-            : explode(',', str_replace(' ', '', strval(env('NO_PROXY')))),
+        'no' => empty(env('NO_PROXY')) ? null : explode(',', str_replace(' ', '', env('NO_PROXY'))),
     ]),
 
     /*
@@ -219,9 +215,7 @@ return [
     |
     */
 
-    'notify_release_stages' => empty(env('BUGSNAG_NOTIFY_RELEASE_STAGES'))
-        ? null
-        : explode(',', str_replace(' ', '', strval(env('BUGSNAG_NOTIFY_RELEASE_STAGES')))),
+    'notify_release_stages' => empty(env('BUGSNAG_NOTIFY_RELEASE_STAGES')) ? null : explode(',', str_replace(' ', '', env('BUGSNAG_NOTIFY_RELEASE_STAGES'))),
 
     /*
     |--------------------------------------------------------------------------
@@ -326,9 +320,7 @@ return [
     |
     */
 
-    'discard_classes' => empty(env('BUGSNAG_DISCARD_CLASSES'))
-        ? null
-        : explode(',', strval(env('BUGSNAG_DISCARD_CLASSES'))),
+    'discard_classes' => empty(env('BUGSNAG_DISCARD_CLASSES')) ? null : explode(',', env('BUGSNAG_DISCARD_CLASSES')),
 
     /*
     |--------------------------------------------------------------------------
@@ -339,9 +331,7 @@ return [
     |
     */
 
-    'redacted_keys' => empty(env('BUGSNAG_REDACTED_KEYS'))
-        ? null
-        : explode(',', strval(env('BUGSNAG_REDACTED_KEYS'))),
+    'redacted_keys' => empty(env('BUGSNAG_REDACTED_KEYS')) ? null : explode(',', env('BUGSNAG_REDACTED_KEYS')),
 
     /*
     |--------------------------------------------------------------------------
@@ -370,7 +360,19 @@ return [
     | The maximum number of breadcrumbs to send with a report.
     |
     | This should be an integer between 0-100 (inclusive).
+    |
     */
 
     'max_breadcrumbs' => null,
+
+    /*
+    |--------------------------------------------------------------------------
+    | Attach hidden context
+    |--------------------------------------------------------------------------
+    |
+    | Whether to attach hidden Context data to events as metadata.
+    |
+    */
+
+    'attach_hidden_context' => env('BUGSNAG_ATTACH_HIDDEN_CONTEXT', false),
 ];
