@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-use App\Http\Controllers\PrometheusMetricsController;
+use App\Http\Controllers\V1\PrometheusMetricsController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', PrometheusMetricsController::class);
