@@ -2,8 +2,6 @@
 
 declare(strict_types=1);
 
-//phpcs:ignorefile
-
 /*
 |--------------------------------------------------------------------------
 | Test Case
@@ -30,7 +28,7 @@ uses(
 |
 */
 
-expect()->extend('toBeOne', fn () => $this->toBe(1));
+// expect()->extend('toBeOne', fn () => $this->toBe(1));
 
 /*
 |--------------------------------------------------------------------------
