@@ -4,11 +4,6 @@ declare(strict_types=1);
 
 namespace Tests;
 
-use Illuminate\Foundation\Testing\LazilyRefreshDatabase;
 use Illuminate\Foundation\Testing\TestCase as BaseTestCase;
 
-abstract class TestCase extends BaseTestCase
-{
-    use CreatesApplication;
-    use LazilyRefreshDatabase;
-}
+abstract class TestCase extends BaseTestCase {}
